feat: implement roadmap fixes — Money class, Circuit Breaker, Decimal migration, test suite
- shared/money.py: Production-grade Money value object with Decimal
  - Rejects float (TypeError) to prevent rounding errors (Bug #4)
  - INR rounding (2dp, ROUND_HALF_UP)
  - GST calculation with CGST/SGST/IGST split (no rounding drift)
  - Invoice validation (prevents zero-amount invoices, Bug #20)
  - Commission calculation utility
  - to_paise()/from_paise() for Razorpay integration

- shared/circuit_breaker.py: Three-state circuit breaker pattern
  - CLOSED → OPEN → HALF_OPEN state machine
  - Fail-fast: rejects immediately when OPEN
  - Automatic recovery: HALF_OPEN probe after timeout
  - Decorator pattern: @circuit_breaker('redis')
  - Context manager: async with breaker.protect()
  - Global registry + metrics for observability

- shared/types/schemas.py: float → Decimal migration
  - OrderCreateRequest.total_amount: float → Decimal
  - CompensationEvent.amount: float → Decimal
  - Prevents IEEE 754 rounding errors in money calculations

- tests/test_critical_fixes.py: Comprehensive test suite
  - Decimal arithmetic tests (float bug demonstration)
  - Money class value object tests
  - GST calculation tests (intra/inter state, 5%/18%)
  - Invoice validation tests (Bug #20 verification)
  - Circuit breaker state transition tests
  - FSSAI license verification tests (Bug #5 verification)
  - GSTIN format validation tests

+++ shared/types/schemas.py
@@ -10,6 +10,7 @@ All models include tenant_id for multi-tenancy support.
 
 from __future__ import annotations
 
+from decimal import Decimal
 from enum import Enum
 from typing import Optional, Dict, Any, List
 from pydantic import BaseModel, Field, ConfigDict
@@ -228,7 +229,7 @@ class OrderCreateRequest(BaseModel):
     kitchen_id: str
     tenant_id: str = Field(default="default")
     items: List[Dict[str, Any]] = Field(default_factory=list)
-    total_amount: Optional[float] = None
+    total_amount: Optional[Decimal] = None
     payment_method: str = "UPI"
     user_tier: UserTier = UserTier.FREE
     requested_time: Optional[str] = None
@@ -360,7 +361,7 @@ class CompensationEvent(BaseModel):
     order_id: str
     tenant_id: str = Field(default="default")
     reason: str  # SHELF_EXPIRED, KITCHEN_FAILURE, PAYMENT_CONFLICT
-    amount: float
+    amount: Decimal  # NEVER float for money — use Decimal
     currency: str = "INR"
     auto_triggered: bool = True
